fix(stock): StockHelper รับ StockLogType ตรง แก้ throw ใน Borrow/Return/Damage/Repair
เปลี่ยน param ApplyStockChangeAsync จาก int itemStatus → StockLogTypeEnum logType
และลบ MapItemStatusToStockLogType ที่ map เฉพาะค่า 1-4 (ItemStatus) ทำให้
Borrow=5/Return=6/Damage=7/Repair=8 throw "ไม่รองรับ ItemStatus: N" ทุกครั้ง

Create POST ใน ItemsController ย้าย map status → InitialXxx มาทำใน controller
เพราะ context ที่ว่าเป็น "รับเข้าครั้งแรก" รู้ได้แค่ที่ call site

+++ Controllers/BorrowHistoriesController.cs
@@ -127,7 +127,7 @@ namespace KlangIT_V3.Controllers
             await StockHelper.ApplyStockChangeAsync(
                 _context,
                 bhVM.ItemId,
-                (int)StockLogTypeEnum.Borrow,
+                StockLogTypeEnum.Borrow,
                 deltaAvailable: -bhVM.Amount,
                 deltaBorrowed:  +bhVM.Amount,
                 deltaDamaged:   0,
@@ -212,7 +212,7 @@ namespace KlangIT_V3.Controllers
             await StockHelper.ApplyStockChangeAsync(
                 _context,
                 bhVM.ItemId,
-                (int)StockLogTypeEnum.Return,
+                StockLogTypeEnum.Return,
                 deltaAvailable: +bhVM.ReturnAmount,
                 deltaBorrowed:  -bhVM.ReturnAmount,
                 deltaDamaged:   0,
